ingest-service/requirements.txt
fastapi
uvicorn
neo4j
python-dotenv
structlog
python-json-logger